refactor(rls): name type alias helper as util
